src/pyproject.toml
[tool.poetry]
name = "telestrainer"
version = "0.1.0"
description = ""
authors = ["Elisei <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Telethon = "^1.25.4"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
mypy = "^0.991"
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.black]
line-length = 120
target_version = ['py310']

[tool.isort]
profile = "django"
lines_after_imports = 2
skip = ".venvs/"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
